Updated registration feature based on review
