+++ package.json
@@ -18,7 +18,6 @@
     "@toolpad/core": "^0.12.0",
     "axios": "^1.7.9",
     "js-cookie": "^3.0.5",
-    "lodash.debounce": "^4.0.8",
     "next": "15.1.6",
     "next-auth": "^4.24.11",
     "react": "^19.0.0",
